add edit-on-gh link
This change required putting some css inline as it looks like there no local assets folder at the moment and the css is being compiled from an external library. It was also necessary to add some -about- text to the -about- page. I copied the text from the about post. This can be updated/removed if necessary.